Модернизация интерфейса INotifications. Увеличено максимальное количество типов уведомлений с 8 до 16. Добавлены новые типы уведомлений - в окне со вкладками и в панели задач.

+++ src/plugins/shortcutmanager/shortcutmanager.h
@@ -53,7 +53,7 @@ private:
 private:
 	bool FAllHidden;
 	bool FTrayHidden;
-	uchar FNotifyHidden;
+	ushort FNotifyHidden;
 	QList< QPointer<QWidget> > FHiddenWidgets;
 };
